Run drush-/drupal-specific install-tasks only when hasDrush=true

+++ fabfile.py
@@ -687,7 +687,7 @@ def drush(drush_command):
 @task
 def install():
   check_config()
-  if env.config['hasDrush'] and env.config['useForDevelopment'] and env.config['supportsInstalls']:
+  if env.config['useForDevelopment'] and env.config['supportsInstalls']:
     if 'database' not in env.config:
       print red('missing database-dictionary in config '+current_config)
       exit(1)
@@ -701,17 +701,18 @@ def install():
       mysql_cmd += 'GRANT ALL PRIVILEGES ON '+o['name']+'.* TO '+o['user']+'@localhost IDENTIFIED BY \''+o['pass']+'\'; FLUSH PRIVILEGES;'
 
       run('mysql -u '+o['user']+' --password='+o['pass']+' -e "'+mysql_cmd+'"')
-      with warn_only():
-        run('chmod u+w '+env.config['siteFolder'])
-        run('chmod u+w '+env.config['siteFolder']+'/settings.php')
-        run('rm -f '+env.config['siteFolder']+'/settings.php.old')
-        run('mv '+env.config['siteFolder']+'/settings.php '+env.config['siteFolder']+'/settings.php.old 2>/dev/null')
+      if env.config['hasDrush']:
+        with warn_only():
+          run('chmod u+w '+env.config['siteFolder'])
+          run('chmod u+w '+env.config['siteFolder']+'/settings.php')
+          run('rm -f '+env.config['siteFolder']+'/settings.php.old')
+          run('mv '+env.config['siteFolder']+'/settings.php '+env.config['siteFolder']+'/settings.php.old 2>/dev/null')
 
-      sites_folder = os.path.basename(env.config['siteFolder'])
-      run_drush('site-install minimal  --sites-subdir='+sites_folder+' --site-name="'+settings['name']+'" --account-name=admin --account-pass=admin --db-url=mysql://' + o['user'] + ':' + o['pass'] + '@localhost/'+o['name'])
+        sites_folder = os.path.basename(env.config['siteFolder'])
+        run_drush('site-install minimal  --sites-subdir='+sites_folder+' --site-name="'+settings['name']+'" --account-name=admin --account-pass=admin --db-url=mysql://' + o['user'] + ':' + o['pass'] + '@localhost/'+o['name'])
 
-      if 'deploymentModule' in settings:
-        run_drush('en -y '+settings['deploymentModule'])
+        if 'deploymentModule' in settings:
+          run_drush('en -y '+settings['deploymentModule'])
 
       reset()
   else:
